Really fix the deployment process!
The previous process broke down in a way that I don't fully understand,
but was related to how the output directory was sometimes updated as
part of a pull request, and sometimes updated by a Travis build job.

I think this process is simpler to understand and easier to debug, and
has the added bonus of not requiring the output directory to be checked
in at all.

I have tested that the general idea works on a separate repo.