Good bug find, thanks Peter:
One problem that I've spotted: the Building Async Microservices workshop is now three and a half hours long, while the Event-Sourced Domain Models in Python workshop that follows it is now only an hour and a half. These were originally two and a half hours each -- could this be fixed please?